refactor into dataset helpers file
